Add #sortItems() to do the sorting
Sorting is done by days till due vs priority level. This took longer
than I want to admit to figure out, but it was a bit of a thinker for
sure on how to exactly do it.

The problem is, and I'm not sure why yet, whenver I use .forEach() or
for..of on the returned sorted array, it places it in the wrong order??
But when accessing each element seperately with index, it gets placed
correctly. Will fix later.